\name{perm.data}
\alias{perm}
\alias{perm3}
\alias{perm4}
\docType{data}
\title{
The partial orders of the attribute patterns for \code{\link{labeling}}
}
\description{
The data matrix of the partial orders of the attribute patterns used for the first labeling method (\code{method="1"}; See \code{\link{labeling}} for details). Available matrices are \code{\link{perm3}} and \code{\link{perm4}} for \eqn{K=3} and \eqn{K=4}, respectively.
}
\usage{
data(perm3)
data(perm4)
}
\format{
\code{\link{perm3}} is a \eqn{48 \times 3} matrix and \code{\link{perm4}} is a \eqn{1680384 \times 3} matrix. 
}


\keyword{datasets}
