\name{sim.Q}
\alias{sim.Q}
\docType{data}
\title{
A complete Q-matrix used to generate \code{\link{sim.dat}}.
}
\description{
The Q-matrix used to generate data (\code{\link{sim.dat}}) in this package.
}
\usage{data(sim.Q)}
\format{
A \eqn{14 \times 3} binary matrix.
}
\details{
It is a binary Q-matrix (Tatsuoka, 1985) associating 14 items and 3 attributes.
}
\references{
  Tatsuoka, K. (1985). A probabilistic model for diagnosing misconceptions in the pattern classification approach. \emph{Journal of Educational Statistics, 12}, 55-73.
  }
\examples{
data(sim.Q)
}
\keyword{datasets}
